+++ examples/godley-lavoie_2007/7_bmw.py
@@ -49,7 +49,7 @@ import fsic
 # parser.
 # 'B' suffix indicates a code change for ease of solution. This applies to
 # Equation 7.15 only.
-script = '''
+script = """
 C_s = C_d                                               # 7.1
 I_s = I_d                                               # 7.2
 N_s = N_d                                               # 7.3
@@ -79,7 +79,7 @@ DA = {delta} * K[-1]                                    # 7.18
 K_T = {kappa} * Y[-1]                                   # 7.19
 I_d = {gamma} * (K_T - K[-1]) + DA                      # 7.20
 r_l = r_l_bar                                           # 7.21
-'''
+"""
 
 symbols = fsic.parse_model(script)
 BMW = fsic.build_model(symbols)
@@ -92,17 +92,17 @@ if __name__ == '__main__':
         range(100),  # Enough periods to reach the stationary state
         alpha_0=25, alpha_1=0.75, alpha_2=0.1,
         delta=0.1, gamma=0.15, kappa=1,
-        pr=1)
+        pr=1)  # fmt: skip
 
     # Set the interest rate
     starting_from_zero.r_l_bar = 0.04
-    starting_from_zero.r_l[0] = starting_from_zero.r_m[0] = starting_from_zero.r_l_bar[0]
+    starting_from_zero.r_l[0] = starting_from_zero.r_m[0] = starting_from_zero.r_l_bar[0]  # fmt: skip
 
     starting_from_zero.solve(max_iter=400)
 
     # Take the results from the last period to be the stationary state
     stationary_state = dict(zip(starting_from_zero.names,
-                                starting_from_zero.values[:, -1]))
+                                starting_from_zero.values[:, -1]))  # fmt: skip
 
     # Run a baseline with these values
     baseline = BMW(range(1945, 2010 + 1), **stationary_state)
@@ -129,9 +129,9 @@ if __name__ == '__main__':
     # Figure 7.1: Evolution of household disposable income and consumption,
     #             following an increase in autonomous consumption expenditures
     axes[0, 0].plot(consumption_scenario.span, consumption_scenario.YD,
-                    label='Household disposable income', color='#33C3F0', linestyle='-')
+                    label='Household disposable income', color='#33C3F0', linestyle='-')  # fmt: skip
     axes[0, 0].plot(consumption_scenario.span, consumption_scenario.C_d,
-                    label='Consumption', color='#FF4F2E', linestyle='--')
+                    label='Consumption', color='#FF4F2E', linestyle='--')  # fmt: skip
 
     axes[0, 0].set_xlim(1950, 2000)
     axes[0, 0].legend()
@@ -140,9 +140,9 @@ if __name__ == '__main__':
     # Figure 7.2: Evolution of gross investment and disposable investment,
     #             following an increase in autonomous consumption expenditures
     axes[0, 1].plot(consumption_scenario.span, consumption_scenario.I_d,
-                    label='Investment', color='#4563F2', linestyle='-')
+                    label='Investment', color='#4563F2', linestyle='-')  # fmt: skip
     axes[0, 1].plot(consumption_scenario.span, consumption_scenario.DA,
-                    label='Depreciation', color='#33C3F0', linestyle='--')
+                    label='Depreciation', color='#33C3F0', linestyle='--')  # fmt: skip
 
     axes[0, 1].set_xlim(1950, 2000)
     axes[0, 1].legend()
@@ -151,7 +151,7 @@ if __name__ == '__main__':
     # Figure 7.4a: Evolution of the output to capital ratio (Y/K−1)
     axes[0, 2].plot(consumption_scenario.span,
                     consumption_scenario.Y / np.hstack([np.nan, consumption_scenario.K[:-1]]),
-                    label='Output:capital ratio', color='#77C3AF', linestyle='-')
+                    label='Output:capital ratio', color='#77C3AF', linestyle='-')  # fmt: skip
 
     axes[0, 2].set_xlim(1950, 2000)
     axes[0, 2].legend()
@@ -159,7 +159,7 @@ if __name__ == '__main__':
 
     # Figure 7.5a: Evolution of the real wage rate (W)
     axes[0, 3].plot(consumption_scenario.span, consumption_scenario.W,
-                    label='Real wage rate', color='#FF992E', linestyle='-')
+                    label='Real wage rate', color='#FF992E', linestyle='-')  # fmt: skip
 
     axes[0, 3].set_xlim(1950, 2000)
     axes[0, 3].legend()
@@ -169,18 +169,18 @@ if __name__ == '__main__':
     #             following an increase in the propensity to save out of
     #             disposable income
     axes[1, 0].plot(saving_scenario.span, saving_scenario.YD,
-                    label='Household disposable income', color='#33C3F0', linestyle='-')
+                    label='Household disposable income', color='#33C3F0', linestyle='-')  # fmt: skip
     axes[1, 0].plot(saving_scenario.span, saving_scenario.C_d,
-                    label='Consumption', color='#FF4F2E', linestyle='--')
+                    label='Consumption', color='#FF4F2E', linestyle='--')  # fmt: skip
 
     axes[1, 0].set_xlim(1950, 2000)
     axes[1, 0].set_title('Figure 7.3: Household disposable income and consumption')
 
     # Figure 7.2a: Evolution of gross investment and disposable investment
     axes[1, 1].plot(saving_scenario.span, saving_scenario.I_d,
-                    label='Investment', color='#4563F2', linestyle='-')
+                    label='Investment', color='#4563F2', linestyle='-')  # fmt: skip
     axes[1, 1].plot(saving_scenario.span, saving_scenario.DA,
-                    label='Depreciation', color='#33C3F0', linestyle='--')
+                    label='Depreciation', color='#33C3F0', linestyle='--')  # fmt: skip
 
     axes[1, 1].set_xlim(1950, 2000)
     axes[1, 1].set_title('Figure 7.2a: Investment and depreciation')
@@ -190,7 +190,7 @@ if __name__ == '__main__':
     #             income
     axes[1, 2].plot(saving_scenario.span,
                     saving_scenario.Y / np.hstack([np.nan, saving_scenario.K[:-1]]),
-                    label='Output:capital ratio', color='#77C3AF', linestyle='-')
+                    label='Output:capital ratio', color='#77C3AF', linestyle='-')  # fmt: skip
 
     axes[1, 2].set_xlim(1950, 2000)
     axes[1, 2].set_title('Figure 7.4: Output:capital ratio')
@@ -198,7 +198,7 @@ if __name__ == '__main__':
     # Figure 7.5: Evolution of the real wage rate (W), following an increase in
     #             the propensity to save out of disposable income
     axes[1, 3].plot(saving_scenario.span, saving_scenario.W,
-                    label='Real wage rate', color='#FF992E', linestyle='-')
+                    label='Real wage rate', color='#FF992E', linestyle='-')  # fmt: skip
 
     axes[1, 3].set_xlim(1950, 2000)
     axes[1, 3].set_title('Figure 7.5: Real wage rate')