+++ fsic/core/containers.py
@@ -6,7 +6,8 @@ Core `VectorContainer` class to handle collections of one-dimensional data.
 import copy
 import difflib
 import re
-from typing import Any, Dict, Hashable, List, Optional, Sequence, Tuple, Union
+from typing import Any, Callable, Dict, Hashable, List, Optional, Sequence, Tuple, Union
+import warnings
 
 import numpy as np
 
@@ -77,7 +78,31 @@ class VectorContainer:
     array([12.,  1., 12.,  3., 12.,  5., 12.,  7., 12.,  9., 10.])
     """
 
-    _VALID_INDEX_METHODS: List[str] = ['get_loc', 'index']
+    @staticmethod
+    def _locate_period_in_span_fallback(period: Hashable, span: np.ndarray) -> int:
+        """Fallback (static) location method, should other `span`-indexing methods fail."""
+        # Convert `span` to a NumPy array of type `object` and locate matches
+        locations = np.asarray(np.asarray(span, dtype=object) == period).nonzero()
+
+        # For now(?), only support one-dimensional array-likes
+        assert len(locations) == 1
+        positions = locations[0]  # Take first (sole) set of axis indexes only
+
+        # No matches: `period` not defined
+        if len(positions) == 0:
+            raise KeyError(period)
+
+        # One match: Unpack and return
+        if len(positions) == 1:
+            return positions[0]
+
+        raise NotImplementedError('Multiple matches not supported')
+
+    _VALID_INDEX_METHODS: List[Union[str, Callable]] = [
+        'get_loc',
+        'index',
+        _locate_period_in_span_fallback,
+    ]
 
     def __init__(self, span: Sequence[Hashable], *, strict: bool = False) -> None:
         """Initialise the container with a defined and labelled `span`.
@@ -285,18 +310,42 @@ class VectorContainer:
         The class-level attribute `_VALID_INDEX_METHODS` defines recognised
         methods for matching.
         """
-        for method in self._VALID_INDEX_METHODS:
-            if hasattr(self.__dict__['span'], method):
-                index_function = getattr(self.__dict__['span'], method)
+        for i, method in enumerate(self._VALID_INDEX_METHODS):
+            if isinstance(method, str):
+                if hasattr(self.__dict__['span'], method):
+                    index_function = getattr(self.__dict__['span'], method)
 
+                    try:
+                        return index_function(period)
+                    except Exception as e:
+                        raise KeyError(period) from e
+
+            elif isinstance(method, Callable):
                 try:
-                    return index_function(period)
+                    return method(period, self.__dict__['span'])
                 except Exception as e:
                     raise KeyError(period) from e
 
+            # This block is for Python versions pre-3.10, in which static
+            # methods weren't callable
+            # The only difference with the block above is the use of the
+            # `__func__` attribute
+            elif isinstance(method, staticmethod):
+                try:
+                    return method.__func__(period, self.__dict__['span'])
+                except Exception as e:
+                    raise KeyError(period) from e
+
+            else:
+                raise TypeError(
+                    f'Unrecognised type ({type(method)}) of search method '
+                    f'with index {i} in `self._VALID_INDEX_METHODS`: {method}'
+                )
+
         raise AttributeError(
             f'Unable to find valid search method in `span`; '
-            f'expected one of: {self._VALID_INDEX_METHODS}'
+            f'expected one of the following, as listed in `self._VALID_INDEX_METHODS`: '
+            f'{self._VALID_INDEX_METHODS}'
         )
 
     def _resolve_period_slice(self, index: slice) -> Tuple[int]:
@@ -709,7 +758,8 @@ class VectorContainer:
         *,
         globals: Optional[Dict[str, Any]] = None,
         locals: Optional[Dict[str, Any]] = None,
-        builtins: [Optional[Dict[str, Any]]] = None,
+        builtins: Optional[Dict[str, Any]] = None,
+        warnings_: str = 'always',
     ) -> Union[float, np.ndarray]:
         """Evaluate `expression` as it applies to the current object. **Uses `eval()`**.
 
@@ -728,6 +778,14 @@ class VectorContainer:
             If `None` (the default), make standard operators from
             `fsic.functions` available (see Notes)
             Disable by passing `{}`
+        warnings_ :
+            Argument to pass to `warnings.simplefilter()` before `eval`uating
+            `expression`. For example, use:
+             - 'ignore' to suppress warnings
+             - 'error' to convert warnings to errors
+
+            See `warnings.simplefilter()` documentation for more information:
+            https://docs.python.org/3/library/warnings.html#describing-warning-filters
 
         Examples
         --------
@@ -850,7 +908,10 @@ class VectorContainer:
         # Either...
         try:
             # ...return the result...
-            return eval(expression, globals, locals_)
+            with warnings.catch_warnings() as w:  # noqa: F841
+                warnings.simplefilter(warnings_)
+
+                return eval(expression, globals, locals_)
 
         except NameError as e:
             # ...or catch a name error (the expression contains an undefined